+++ NetCoreWebApiPoC.Domain/Exceptions/BadRequestException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace NetCoreWebApiPoC.Data
+namespace NetCoreWebApiPoC.Domain.Exceptions
 {
     public class BadRequestException : Exception
     {
