+++ NetCoreWebApiPoC.Domain/Exceptions/NotFoundException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace NetCoreWebApiPoC.Data
+namespace NetCoreWebApiPoC.Domain.Exceptions
 {
     public class NotFoundException : Exception
     {
